update suites after AiPersonaService removal
