+++ lib/src/scenario/graphql/queries/virtual-contributor/getAiPersonaModelCard.graphql
@@ -1,28 +1,25 @@
 query GetAiPersonaModelCard($id: UUID!) {
   lookup {
     virtualContributor(ID: $id) {
-      aiPersona {
-        id
-        modelCard {
-          spaceUsage {
-            modelCardEntry
-            flags {
-              name
-              enabled
-            }
+      modelCard {
+        spaceUsage {
+          modelCardEntry
+          flags {
+            name
+            enabled
           }
-          aiEngine {
-            isExternal
-            hostingLocation
-            isUsingOpenWeightsModel
-            isInteractionDataUsedForTraining
-            canAccessWebWhenAnswering
-            areAnswersRestrictedToBodyOfKnowledge
-            additionalTechnicalDetails
-          }
-          monitoring {
-            isUsageMonitoredByAlkemio
-          }
+        }
+        aiEngine {
+          isExternal
+          hostingLocation
+          isUsingOpenWeightsModel
+          isInteractionDataUsedForTraining
+          canAccessWebWhenAnswering
+          areAnswersRestrictedToBodyOfKnowledge
+          additionalTechnicalDetails
+        }
+        monitoring {
+          isUsageMonitoredByAlkemio
         }
       }
     }
