+++ lib/src/scenario/graphql/queries/virtual-contributor/virtualContributor.graphql
@@ -67,13 +67,11 @@ query VirtualContributor($id: UUID!) {
       searchVisibility
       listedInStore
       status
+      bodyOfKnowledgeID
+      bodyOfKnowledgeType
+      bodyOfKnowledgeDescription
       aiPersona {
-        id
         engine
-        bodyOfKnowledgeID
-        bodyOfKnowledgeType
-        bodyOfKnowledge
-        __typename
       }
       profile {
         id
