+++ server-api/src/functional-api/contributor-management/virtual-contributor/model-card/model-card.request.params.ts
@@ -1,7 +1,7 @@
 import { getGraphqlClient, TestUser } from '@alkemio/tests-lib';
 import {
-  AiPersonaBodyOfKnowledgeType,
   AiPersonaEngine,
+  VirtualContributorBodyOfKnowledgeType,
 } from '@alkemio/tests-lib/core/generated/alkemio-schema';
 import { graphqlErrorWrapper } from '@alkemio/tests-lib/utils/graphql.wrapper';
 
@@ -54,12 +54,10 @@ export const createVirtualContributorWithEngineType = async (
             displayName,
           },
           accountID,
+          bodyOfKnowledgeType:
+            VirtualContributorBodyOfKnowledgeType.AlkemioKnowledgeBase,
           aiPersona: {
-            aiPersonaService: {
-              bodyOfKnowledgeType:
-                AiPersonaBodyOfKnowledgeType.AlkemioKnowledgeBase,
-              engine: engineType,
-            },
+            engine: engineType,
           },
         },
       },
